% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{class}
\name{connection}
\alias{connection}
\alias{dbConnect,RedashMySQLDriver-method}
\title{Connection for 'Redash' 'MySQL' Database}
\usage{
\S4method{dbConnect}{RedashMySQLDriver}(drv, ...)
}
\arguments{
\item{drv}{An object created by \code{RedashMySQL}}
}
\description{
Connection for 'Redash' 'MySQL' Database
}
\examples{
\dontrun{
db <- dbConnect(RedashMySQL::RedashMySQL())
dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
\keyword{internal}
