% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\docType{class}
\name{driver}
\alias{driver}
\alias{dbUnloadDriver,RedashDriver-method}
\title{Driver for 'Redash' Database}
\usage{
\S4method{dbUnloadDriver}{RedashDriver}(drv, ...)
}
\description{
Driver for 'Redash' Database
}
\keyword{internal}
