% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedashMySQL-package.R
\docType{package}
\name{RedashMySQL}
\alias{RedashMySQL}
\alias{RedashMySQL-package}
\title{Database Interface and 'MySQL' Driver via Redash}
\description{
Database Interface and 'MySQL' Driver via Redash
}
