% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{class}
\name{connection}
\alias{connection}
\alias{dbConnect,RedashDriver-method}
\title{Connection for 'Redash' Database}
\usage{
\S4method{dbConnect}{RedashDriver}(drv, ...)
}
\arguments{
\item{drv}{An object created by \code{Redash}}
}
\description{
Connection for 'Redash' Database
}
\examples{
\dontrun{
db <- dbConnect(Redashr::Redash())
dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
\keyword{internal}
