% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\docType{class}
\name{driver}
\alias{driver}
\alias{dbUnloadDriver,RedashMySQLDriver-method}
\title{Driver for 'Redash' 'MySQL' Database}
\usage{
\S4method{dbUnloadDriver}{RedashMySQLDriver}(drv, ...)
}
\description{
Driver for 'Redash' 'MySQL' Database
}
\keyword{internal}
